Autocompletion of magic comment values
